+++ walletd-staking/src/cosmos.rs
@@ -11,8 +11,8 @@ impl Default for CosmosStaking { fn default() -> Self { Self::new() } }
 impl Staking for CosmosStaking {
     async fn get_stake_info(&self, _address: &str) -> StakingResult<StakeInfo> { Ok(StakeInfo { total_staked: 0, available_balance: 0, pending_rewards: 0, delegations: Vec::new(), unbonding: Vec::new() }) }
     async fn get_validators(&self, limit: usize) -> StakingResult<Vec<ValidatorInfo>> { Ok((0..limit.min(10)).map(|i| ValidatorInfo { address: format!("cosmosvaloper1{:038}", i), name: format!("Chorus One {}", i), commission: 5.0, total_stake: 1_000_000_000, delegators: 300, uptime: 99.9, apy: 18.0, is_active: true }).collect()) }
-    async fn stake(&self, amount: u64, _validator: &str) -> StakingResult<Vec<u8>> { if amount == 0 { return Err(StakingError::InvalidAmount); } Ok(vec![0; 32]) }
-    async fn unstake(&self, _amount: u64, _validator: &str) -> StakingResult<Vec<u8>> { Ok(vec![0; 32]) }
+    async fn stake(&self, amount: u128, _validator: &str) -> StakingResult<Vec<u8>> { if amount == 0 { return Err(StakingError::InvalidAmount); } Ok(vec![0; 32]) }
+    async fn unstake(&self, _amount: u128, _validator: &str) -> StakingResult<Vec<u8>> { Ok(vec![0; 32]) }
     async fn claim_rewards(&self) -> StakingResult<Vec<u8>> { Ok(vec![0; 32]) }
     async fn get_apy(&self) -> StakingResult<f64> { Ok(18.0) }
 }
